fix(ui): relocate LevelBarsButton to patterns, fixing the only primitives->patterns dependency
LevelBarsButton imported ComposerControlButton from ../patterns/, the
tree's sole primitives->patterns edge and a direct violation of the
tier rule (primitives compose nothing above their own tier). Move the
file to patterns/, fix its now-sibling import to ./ComposerControlButton,
repoint the package.json export key, the product-client consumer, and
the test import, and move its component-library.md row from Primitives
to Patterns.
